# src/docs/openapi.generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API RRHH Hospital - Personal",
    "version": "1.0.0",
    "description": "OpenAPI enterprise (auto)"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Personal": {
        "type": "object",
        "properties": {
          "dni": {
            "type": "string"
          },
          "apellido": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "fecha_nacimiento": {
            "type": "string"
          },
          "sexo_id": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "domicilio": {
            "type": "string"
          },
          "localidad_id": {
            "type": "string"
          },
          "cuil": {
            "type": "string"
          },
          "nacionalidad": {
            "type": "string"
          },
          "observaciones": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
      "Tareas": {
        "type": "object",
        "properties": {
          "nombre_tarea": {
            "type": "string"
          },
          "descripcion_tarea": {
            "type": "string"
          },
          "estado": {
            "type": "string"
          },
          "fecha": {
            "type": "string"
          },
          "asignado_a": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/personal": {
      "get": {
        "tags": [
          "personal"
        ],
        "summary": "Listar personal",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "id"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ],
              "default": "ASC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "personal"
        ],
        "summary": "Crear personal",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Personal"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creado"
          }
        }
      }
    },
    "/api/personal/{id}": {
      "get": {
        "tags": [
          "personal"
        ],
        "summary": "Obtener personal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "personal"
        ],
        "summary": "Actualizar personal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Personal"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "personal"
        ],
        "summary": "Eliminar personal (soft-delete)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Borrado"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      }
    },
...
    "/api/tareas": {
      "get": {
        "tags": [
          "tareas"
        ],
        "summary": "Listar tareas",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "id"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ],
              "default": "ASC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "tareas"
        ],
        "summary": "Crear tareas",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Tareas"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creado"
          }
        }
      }
    },
    "/api/tareas/{id}": {
      "get": {
        "tags": [
          "tareas"
        ],
        "summary": "Obtener tareas",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "tareas"
        ],
        "summary": "Actualizar tareas",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Tareas"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "tareas"
        ],
        "summary": "Eliminar tareas (soft-delete)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Borrado"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      }
    },
...
  }
}